refactor(deploy): split checkWorkbenchAppDir out of checkDir
A boolean option made checkDir's contract depend on the app shape.
Callers already know whether they deploy a workbench app, so each build
shape gets its own self-contained check and checkDir keeps its original
index.html contract.
