+++ packages/@sanity/cli/src/actions/deploy/__tests__/checkDir.test.ts
@@ -1,4 +1,4 @@
-import {readFile, stat} from 'node:fs/promises'
+import {stat} from 'node:fs/promises'
 import {join} from 'node:path'
 
 import {afterEach, describe, expect, test, vi} from 'vitest'
@@ -6,24 +6,10 @@ import {afterEach, describe, expect, test, vi} from 'vitest'
 import {checkDir} from '../checkDir.js'
 
 vi.mock('node:fs/promises', () => ({
-  readFile: vi.fn(),
   stat: vi.fn(),
 }))
 
 const mockStat = vi.mocked(stat)
-const mockReadFile = vi.mocked(readFile)
-
-const enoent = () => {
-  const error = new Error('ENOENT') as NodeJS.ErrnoException
-  error.code = 'ENOENT'
-  return error
-}
-
-const mockSourceDirExists = () => {
-  mockStat.mockResolvedValueOnce({
-    isDirectory: () => true,
-  } as never)
-}
 
 describe('#checkDir', () => {
   const testDir = '/test/directory'
@@ -170,72 +156,5 @@ describe('#checkDir', () => {
     expect(mockStat).toHaveBeenCalledTimes(2)
     expect(mockStat).toHaveBeenNthCalledWith(1, testDir)
     expect(mockStat).toHaveBeenNthCalledWith(2, join(testDir, 'index.html'))
-  })
-
-  describe('workbench (federation) builds', () => {
-    const manifestPath = join(testDir, 'mf-manifest.json')
-
-    const validManifest = {
-      exposes: [
-        {
-          assets: {
-            css: {async: [], sync: ['assets/panel.css']},
-            js: {async: [], sync: ['assets/panel.js']},
-          },
-          name: 'views/favorites/panel',
-        },
-      ],
-      metaData: {
-        remoteEntry: {name: 'remote-entry.js'},
-      },
-    }
-
-    test('should pass when the manifest exists and exposes at least one module', async () => {
-      mockSourceDirExists()
-      mockReadFile.mockResolvedValueOnce(JSON.stringify(validManifest))
-
-      await expect(checkDir(testDir, {isWorkbenchApp: true})).resolves.toBeUndefined()
-
-      expect(mockReadFile).toHaveBeenCalledWith(manifestPath, 'utf8')
-      // never falls through to the index.html check
-      expect(mockStat).not.toHaveBeenCalledWith(join(testDir, 'index.html'))
-    })
-
-    test('should throw when mf-manifest.json does not exist', async () => {
-      mockSourceDirExists()
-      mockReadFile.mockRejectedValueOnce(enoent())
-
-      await expect(checkDir(testDir, {isWorkbenchApp: true})).rejects.toThrow(
-        `"${manifestPath}" does not exist`,
-      )
-    })
-
-    test('should throw when mf-manifest.json is not valid JSON', async () => {
-      mockSourceDirExists()
-      mockReadFile.mockResolvedValueOnce('not json{')
-
-      await expect(checkDir(testDir, {isWorkbenchApp: true})).rejects.toThrow(
-        `"${manifestPath}" is not valid JSON`,
-      )
-    })
-
-    test('should throw when manifest declares no exposed modules', async () => {
-      mockSourceDirExists()
-      mockReadFile.mockResolvedValueOnce(JSON.stringify({...validManifest, exposes: []}))
-
-      await expect(checkDir(testDir, {isWorkbenchApp: true})).rejects.toThrow(
-        'declares no exposed modules',
-      )
-    })
-
-    test('should re-throw non-ENOENT errors when reading the manifest', async () => {
-      mockSourceDirExists()
-
-      const permissionError = new Error('Permission denied') as NodeJS.ErrnoException
-      permissionError.code = 'EACCES'
-      mockReadFile.mockRejectedValueOnce(permissionError)
-
-      await expect(checkDir(testDir, {isWorkbenchApp: true})).rejects.toThrow('Permission denied')
-    })
   })
 })
